+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - set background opacity for kitty
 - add fcitx5
+- set theme for kitty
 
 ## 2021-12-17
 ### Added